cleanup and fixes
